feat: mutualize getPReferredName
